perf(fe-zerocopy): dtype passthrough + ndarray buffer in FE transformers + cache key fix
* training/core/_phase_recurrent.py: extend _coerce_features_to_float32 cache key from (split, id(frame)) to (split, id(frame), cols_signature) so an in-place column add/remove on the same frame id no longer serves a stale cached ndarray with the wrong column layout.
* feature_engineering/bruteforce.py: _kfold_target_encode pre-allocates an ndarray buffer and wraps to DataFrame once at return, avoiding per-fold DataFrame.iloc[val_idx, :] = ... block-consolidation copies (measured 1.46x on N=10000 K=5 NC=10).
* feature_engineering/transformer/stacked_attention.py + boosted_attention.py: forward caller dtype to compute_row_attention so the inner ndarray is produced at the requested dtype directly; without this the inner default float32 forced a full-buffer float32 -> float64 copy per layer when caller dtype was float64.
* feature_engineering/transformer/per_column_rff.py + random_features.py: documented bench-attempt-rejected on the polars allow_copy=False hint (raises unconditionally on any multi-column DataFrame; per-column Arrow chunk layout). Existing astype(dtype, copy=False) gate is already correct.
* feature_engineering/transformer/stacked_qnn.py: documented why Q1 layer stays float32 by design.
* feature_engineering/timeseries.py: documented why dtype= must stay on the constructor (list-of-rows input has no .dtype).
* feature_engineering/bruteforce.py: documented bench-attempt-rejected on use_pyarrow_extension_array=True (measured 3.0x slower on the PySR -> sklearn.check_array path; pyarrow extension dtypes densify back to float64 ndarray and add an extra Arrow -> numpy round-trip).
* tests/feature_engineering/transformer/test_regression_w2c_dtype_gate.py: 8 sensors covering cache-key invalidation, kfold buffer behaviour, and dtype passthrough on stacked / boosted attention.

+++ src/mlframe/feature_engineering/transformer/stacked_attention.py
@@ -82,13 +82,14 @@ def compute_stacked_row_attention(
         layer_seed = seed + 1000 * layer
         # head_dim for this layer cannot exceed current_X_train.shape[1]; clamp adaptively (later layers have low-d input).
         layer_head_dim = min(head_dim, max(2, current_X_train.shape[1] - 1))
-        # Mode A: OOF y_mean for X_train.
+        # Mode A: OOF y_mean for X_train. Pass caller dtype through so .to_numpy() yields the requested dtype directly; the trailing astype(..., copy=False) is
+        # then a no-op identity. Without dtype= the inner default is float32 -- callers using float64 paid a full-buffer float32->float64 copy per layer.
         out_train = compute_row_attention(
             X_train=current_X_train, y_train=y_train, X_query=None, splitter=splitter,
             seed=layer_seed, n_heads=n_heads, head_dim=layer_head_dim, k=k,
             softmax_temp=softmax_temp, aggregate=("y_mean",), standardize=standardize,
             projection=projection, gpu_stage4=False, dedupe_threshold=None,
-            allow_overcomplete=True,
+            allow_overcomplete=True, dtype=dtype,
         ).to_numpy().astype(dtype, copy=False)
         layer_outputs_train.append(out_train)
 
@@ -99,7 +100,7 @@ def compute_stacked_row_attention(
                 seed=layer_seed, n_heads=n_heads, head_dim=layer_head_dim, k=k,
                 softmax_temp=softmax_temp, aggregate=("y_mean",), standardize=standardize,
                 projection=projection, gpu_stage4=False, dedupe_threshold=None,
-                allow_overcomplete=True,
+                allow_overcomplete=True, dtype=dtype,
             ).to_numpy().astype(dtype, copy=False)
             layer_outputs_query.append(out_query)
